Remove duplicate form component and update imports
