+++ components/contract-generator-form.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, waitFor } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
-import GenerateContractForm from "./GenerateContractForm"
+import GenerateContractForm from "./generate-contract-form"
 
 const toastMock = jest.fn()
 jest.mock("@/hooks/use-toast", () => ({
